improve bfloat16 numerical stability: SDPA and float32 loss
1. Replace manual softmax attention with F.scaled_dot_product_attention
   which internally computes in float32 for bfloat16 inputs and enables
   FlashAttention/memory-efficient kernels when available.
2. Cast logits to float32 before CrossEntropyLoss to avoid precision
   loss in log_softmax computation.

Both changes are weight-compatible with existing checkpoints.

+++ recursive_compressor_lm.py
@@ -35,7 +35,7 @@ class RecursiveCompressorLM(PreTrainedModel):
         logits, _ = self.step(input_ids, None)
         loss = None
         if labels is not None:
-            loss = nn.CrossEntropyLoss()(logits.view(-1, self.config.vocab_size), labels.view(-1))
+            loss = nn.CrossEntropyLoss()(logits.float().view(-1, self.config.vocab_size), labels.view(-1))
         return CausalLMOutput(loss=loss, logits=logits)
 
     def predict(self, input_ids, hidden):
